// packages/components/src/image/editor/components/CropprWrapper/ArchbaseCropperWrapper.d.ts
import React from 'react';
import './ArchbaseCropper.scss';
export declare const ArchbaseCropperWrapper: React.MemoExoticComponent<({ src, size, croppUpdate }: {
    src: string;
    size: any;
    croppUpdate: (data: {
        x: number;
        y: number;
        width: number;
        height: number;
    }) => void;
}) => import("react/jsx-runtime").JSX.Element>;
//# sourceMappingURL=ArchbaseCropperWrapper.d.ts.map